+++ app/controllers/web/admin/services_controller.rb
@@ -1,14 +1,17 @@
 class Web::Admin::ServicesController < Web::Admin::ApplicationController
   def index
-    @services = Service.all
+    @services = Service.all.page(params[:page])
+    authorize current_user, policy_class: AdminPolicy
   end
 
   def new
     @service = Service.new
+    authorize current_user, policy_class: AdminPolicy
   end
 
   def create
     @service = Service.new(service_attrs)
+    authorize current_user, policy_class: AdminPolicy
     if @service.save
       redirect_to action: :index
     else
@@ -18,14 +21,17 @@ class Web::Admin::ServicesController < Web::Admin::ApplicationController
 
   def show
     @service = Service.find(params[:id])
+    authorize current_user, policy_class: ClientPolicy
   end
 
   def edit
     @service = Service.find(params[:id])
+    authorize current_user, policy_class: ClientPolicy
   end
 
   def update
     @service = Service.find(params[:id])
+    authorize current_user, policy_class: ClientPolicy
     if @service.update(service_attrs)
       redirect_to action: :index
     else
@@ -34,6 +40,7 @@ class Web::Admin::ServicesController < Web::Admin::ApplicationController
   end
 
   def destroy
+    authorize current_user, policy_class: AdminPolicy
     service = Service.find(params[:id])
     service.destroy
